requirements.txt
dockerfile-parse
docker
PyYAML
